Created new load_util module with multiprocessing logic to avoid repeated imports of mantid
